+++ features/introduction/src/main/java/com/diploma/work/introduction/IntroductionScreen.kt
@@ -33,7 +33,6 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.deploma.work.introduction.R
-import com.diploma.work.features.MainButton
 import com.diploma.work.features.ScreenRoute.HOME
 import com.diploma.work.features.ScreenRoute.INTRODUCTION
 import kotlinx.coroutines.launch
@@ -131,7 +130,7 @@ fun IntroductionScreen(
             }
             Row(modifier = Modifier.fillMaxWidth()) {
                 if (pagerState.currentPage > 0) {
-                    MainButton(
+                    com.diploma.work.composables.MainButton(
                         modifier = Modifier
                             .padding(8.dp)
                             .wrapContentWidth()
@@ -153,7 +152,7 @@ fun IntroductionScreen(
                         },
                     )
                 }
-                MainButton(
+                com.diploma.work.composables.MainButton(
                     modifier = Modifier
                         .padding(8.dp)
                         .wrapContentWidth()
