+++ settings.gradle.kts
@@ -23,3 +23,4 @@ include(":core")
 include(":core:database")
 include(":core:sharedpreference")
 include(":core:repository")
+include(":core:composables")
